pyproject.toml
[tool.poetry]
name = "homework-30"
version = "0.1.0"
description = ""
authors = ["TatianaBelova333 <101121235+TatianaBelova333@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "homework_30"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
psycopg2 = "^2.9.3"
Pillow = "^9.2.0"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
